requirements.txt
customtkinter
pillow
fastapi
uvicorn
sqlalchemy
psycopg2-binary